requirements.txt
asgiref==3.7.2
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Faker==21.0.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
psycopg==3.1.14
PyJWT==2.8.0
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
uuid==1.30
